requirements.txt
yt-dlp==2025.9.5
Django==5.2
djangorestframework==3.16.1